Initial commit of second version.
Copied over files from first version.
New version better incorporates cpp code due to ADzemski (rtnorm).
Put test of the code in a vignette in which we try to use the code to reproduce tables similar to those in the original 1996 publication.
This motivated the development of a few additional functions to aid in plotting or reproduce functions and concepts from the paper (Law and Morton, 1996).