Testing Permanence method.
Created permanence vignette, which is almost a reproduction of the numerical vignette. (Still need to compare results.)

Major:
Added easy extraction of final equilibrium.
Fixed a comment.
Fixed adding a nonzero density to represent a new arrival.
Switched from an arbitrary large point to the previous solution of the steady-state evaluation or of the parent community.
Switched from solving a steady-state system (which tends to move towards a zero population).
Adjusted rounding precision to fix some edge cases where a system was apparently not permanent but all of its subcommunities that were permanent were invadable by it.
Added handling of the case where two subcommunities are uninvadable.

Minor:
Added some safeties to the "addMissingEquilibriaEntries" function.
Fixed a typo in my Table 1 matrix.
Began working on Table 7 (Invasion Resistance).