Started working on Scaling document/experiment.
Added todos as well.
Note that we will need to probably turn this into a generate-save-clusterRun-load-analyse workflow.
Hence, definitely need to take care fo the parallelisation todo ASAP.
And then there is the possibility of checking how the scalings respond to the parameters, for which we can probably dial back down the largest runs (3000 species in the pool, probably would only need the 90~ case if the scaling is nice).