Working on Islands. Still need to finish and test I believe.
Added a comment to the arrival density in Law and Morton Numerical Assembly, emphasising the importance of the arrival density. Since we work in the continuum, the arrival density is an approximation for epsilon > 0, but in discrete space this is somewhat absurd and can cause instabilities. One of our example cases changes the fixed point it travels to if the arrival density is greater than about 0.3. Indeed, this drives the result we see in the Islands context.

Created a second version of the island example runs.

Fixed an apparently minor issues in the parantheses in the parallel version of the code.