Two major contributions: first to analysing existing Viking data, second to generating some more viking data.
First, see CommunityIslands, TestRun2, and the Rmd notebook.
These have new functions and attempts to try to find the abundances that I did not include in the first version of the Viking code (but which can be seen in the new version of the viking code, not included here).
The chosen result is stored in the notebook and essentially entails using all Type 2 and Type 3 (i.e. something happened) invasions to re-arrive at the result.