Discovered issue with the 2021-04 data in that it did not include abundances. Created 2021-05 data with abundances to compare with.
Essentially finished the first draft of the scaling results. I do not think there is (yet) a publication here, if only because I think other models should be compared with. This goes doubly so for more modern models.

Island results seem acceptable, but I am surprised not to see more stalemates, like I had seen initially. It might be good to see why this is.