Running the diffusion data through some of the existing code to see what happens.
Early looks suggest that it gives more or less (a subset of) the same results.

It is worth noting that there appears to be at least one bug (a species is repeated inside of the community listing).

More concerning, some of the assembled states are being listed as not uninvadable. This will need to be checked at some stage, but should be a problem limited to a few systems which were initially discarded due to a lack of a recorded steady-state.