Testing the Numerical Assembly function.
Most of it is a straight copy of the LM1996 equivalent, but Run_GLV seems to be stabler so far.
The basic dynamics seem to work well, but there is something wrong with the invadability checking mechanism I added.
This is important in the case of carrying capacity dynamics, where there is a frequent "edge" case, where a microbe dies so quickly that another abundance skyrockets and triggers the normalisation, which induces positive feedback.