Identified bug in DynamicsNormalisedCapacity. Checking invadability before invasion opens up quite the can of worms though.
So, if invadability is not checked before invasion, which one might regard as a sort of default behaviour, then the size of the invader can matter and we do not act as if the invasion occurs from an infinitesimal.
This violates the thinking of Law and Morton, 1996.
On the other hand, if we do check the behaviour, then the results of the dynamics can change from when we do not.
With normal dynamics, if the propagule pressure is sufficiently high, then a species that was making the system inhospitable can go extinct via competition.
With normalised/carrying capacity dynamics, then the mutualisms/exploitation can drive the invader and the mutualists to outcompete for space the invader's competitors.
In essence, the system becomes highly sensitive to propagule pressure/arrival density in the carrying capacity case.