Realised a subtle error that only showed itself explicitly when I was doing the diffusion cases.
In the diffusion cases, I noticed a species was repeated in the naming convention, suggesting that I had applied the wrong set of names.
In practice, I had applied the first set of names to each combination.
I corrected this, and noted it made a substantial difference in the Questions document, indicating carrythrough errors.
I decided to implement the same system in the earlier version and also noticed substantial discrepancies.
I next will be error checking to make sure the new system works as I anticipate and, if all goes well, I will need to rerun all of my analyses.