Chris's point about using the wrong trophic level calculation.
We should be using the FlowBasedTrophicLevel, but we need to be careful to use the weight.by argument (which the plotwebbylevel function does not recognise due to a coding error).
This might be why we didn't register it before, but that does not forgive the lack of usage of the prey-averaged food web. Oh well.