Added abundance and relative abundance comparisons, although it is not perhaps obvious how they should be implemented.
The most proper way to do so is probably use BiodiversityR with environmental data identified pool and community and "both" to recreate the plots that I have already supplied.

The most proper way to see how a pool-world-system might work is probably to set up source islands for each community assembled, then a colony for each other community assembled, which are then pairwise connected to a neutral island.
Then the full rank-abundance can be calculated on the system (presumably at steady-state.

This might look like, for three systems (o, x, z):
  o--[]--x-x-x--[]--z
o<                   >z
  o--[]---------[]--z
so that there are an o-[]-x, an x-[]-z, and an o-[]-z systems.