Something not quite right with the main function.
It suggests that we are applying Real to a list instead of a numeric internally inside of lsodar, but the check functions appear to work fine.
We might need to check the examples, but our work so far seems consistent with them...

Otherwise, fixed some minor bugs as well as the event timings, which were an unrealised mess.