Code probably will not work as is.
Trying to get code working, but it appears that deSolve's func argument is not allowed to return non-numeric arguments, despite returning the numeric arguments in the first entry completely separately from anything in the other entries.

As such, the conclusion would seem to be to give up on the active fitting of the trophic structure at the same time as the dynamics.
The result is both a lot of memory usage and even more wranging with the code to circumvent the above restriction.
The next obvious step would be to try to get it working by doing it post facto.