Fixing a random correlation in my interaction matrices.
Attempting a potentially controversial but important
change.
I hope that this won't affect any outputs.
If it does, the exact who/what/when/how of it would
be interesting, but devious to diagnose.
I don't have any reason why it would, but it is
worth noting.
Still need to test the results at this stage.

Signed-off-by: Brennen-Fagan <92510995+Brennen-Fagan@users.noreply.github.com>