Trying to test characteristic rate argument.
Realised that we have not tested properly some of the cases.
Note, for instance, that we have already supplied a matrix to the dynamics in order to establish the dynamics properly.
But then we have a new matrix computed in order to compute a time scale.
But that time scale has nothing to do with the actual time scale of the dynamics, because it is a new independent matrix.
The most natural thing seems to be to deprecate this and fix it.