Catching up on Git commits III.
Caught a huge problem here. I had forgotten to include the Inf spatial distance when updating the code before Christmas.

As a result, we have to add it onto the end.

Does this result in problems for our previous runs?

Some of our documentation is wrong and will need to be fixed in the final draft of the code.

The data analysis turned out okay due to a mismatch in a rare occurrence of two wrongs making a right.
All of the last spatial type were neglected, but all of the previous spatial types were calculated as listed here.