Catching up on Git commits IV.
Another problem; quite a few of the simulations were not able to finish due to time and memory constraints.

Another another problem: the solver was causing an out of memory error in uncommon but consistent ways.

To circumvent/solve these problems, we ran each history separately in two parts, saving each part as we go.

This has been largely successful on Viking.