Catching up on Git commits V.
An earlier version of how we were planning on handling the output.

The central idea is to reduce the sampling interval (preferred_rows_per_event) and then, on the reduced set, calculate the three diversity metrics and combine them with the system modifiers.

Note the bug in line 221 that is fixed in a future version, as well as the inability to cope with separate histories (as noted in a previous commit in this batch of commits).