Catching up on Git commits VI.
Two variants, neither up to date, of the same file.

Both files are designed to handle the diversity outputs, which were summarised from the simulation outputs.
In turn, these files produce plots.

At this time, the parameters did not get matched correctly to the data.

A significant portion is combining the data across diversity outputs. We then proceed to attempt multiple plotting strategies.

Additions in the second file include averaging, improvements in the selection (so as to include pool and noise) and more plots.