Beginning merge with research server.
Something went wrong with the version of the project on the research server, so I am not able to do a git merge.

(More accurately, it looks like git was never initialised on that end, so I would need to start over from the top.)

As such, I am manually merging files.

I choose to ignore images where I can, as well as .RData files for file size reasons.
This is fairly restrictive, in that a lot of the work on the server relied on loading up files and epxploring them to decide on the best way to proceed.

Finally, the additional file in this commit is "TempFile-IndepTest.R".
The goal of this file was to test the impact of binning time steps and averaging the result (median) on the correlation between runs.

Correlation is one of the things that we have not quite solved, but it is not clear if it is completely removable and it is not clear if it is an issue.

The latter strikes me in particular since the "correlation" arises from otherwise completely independent runs, up to and including runs from different pools, environments, time steps, and computer.

Regarding the former, this type of correlation is considered "spurious", and arises out of the coupling with time. As I recall, scaling, centring, and removing autocorrelation with an AR(6) process seemed to remove most of the spurious correlation from the system.