Merge with Research Server IV.
Updated versions of some of diversity handling code.
There's a lot going on here.

(Original file)
We've added in further grouping and averaging to try to improve run times while not harming the data or output.

(Version 2)
We've added in tracking of case (1 - 3) and "casenumber" (the id used to select which replicate of which parameter combination), as well as an initial attempt at a 3 x 3 plot, usage of boxplots with means, usage of trajectory plots over heat maps, and an example plot for showcasing to the Leverhulme Trust.
I note that the code should be fine, but due to the off-by-one error and the missing files, not all of these are fully resolved in publishable quality yet.

(v2 Subset file)
An intermediate of the original and v2 used for testing and demonstrating how we might filter down the systems examined for speed of analysis.

(v3)
Taking lessons learned from the diversity code and the above versions,
we rewrite some of the opening lines to accomodate and note the differences in format during loading.
Plots remain essentially the same.

(Viking v2)
A variant of v2 from an earlier stage that attempts to add saving of the plots automatically. This is not up to date with version 3.