Merge with Research Server VI.
Another set of long stories that are out of order with some of the previously recorded commits due to my poor commit history on the server.

The base file here is FirstAttempt-Doc-Analysis2.Rmd.
Edits here include the addition of a burn-in period, some additional safety (line 332), the addition of Jaccard Distance/Beta diversity, an early attempt at species presence in each environment binned by times, better treatment of the distances, significant reformatting of some images, additions of species information plots, and an attempt at dealing with events that didn't go over well.

Most, but not all of these changes made it over into Gallery 1.