Merge with Research Server VII.
(Alternate Version of Analysis2)
Removes the fitdistrplus and poweRlaw components where we attempted to fit the inter event arrival times.
It also originates before Jaccard diversity was brought to my attention.
This resulted in a much tighter script overall which meant less to load.

(Gallery 2)
Another branch.
Explicitly named gallery, added in some notes to myself, but does not have the burn in.
This version also made an attempt at trophic functionality, but the computation proved too complex at the time and has not been attempted since.
The goal had been to try and examine how the trophic structure was changing over time, and too many of the network were too complicated for the cheddar package to finish computing.
This also branched before Jaccard distance was included.