Merge with Research Server IX.
This one is a bit of a nightmare, but was already alluded to, where-in we have a bit of a problem with spurious correlations of time series.
These files attempt to dig into what forms this spurious correlation takes on.

testcorr:
Format an alpha diversity, burn-in, burn-out, demean, and remove autocorrelation and then examine the correlations.
The result still has some problems with autocorrelation as I recall.
In order to run this file, one needs to have run a chunk like chunk 7 (line 425) of Gallery 3.

Alternatively, one might be able to get away with running

dplyr::bind_rows(lapply(1:10, function(i) Diversity[[i]]$alpha %>% dplyr::mutate(
    Simulation = i, Space = "NA", Distance = Diversity$SpaceMod, Modifier = "NA", ModIntensity = toString(Diversity$NeutralMod), Measurement = "Richness", Value = Richness)
)) -> Diversity

on a Diversity output.

When you run it, generally, you see that there is quite a spread of correlation that is significant.
It better controlled cases, this tends to come out symmetric as well.
There is a fair mixture of support for stationarity and non-stationarity, making things more difficult.
We do manage to eliminate most of the partial autocorrelations.
Note lines 101 to the end will need some patching if the entire data set is not considered.
They also show that there does not appear to be a consistent pattern to the correlations even between time slices.