Merge with Research Server X(2).
The other file before my train of thought was interrupted.

This file attempted to look at the Gelman Rubin diagnostic to see if it might suggest something.
For a lot of simulations this was, if not reassuring, then at least neutral.
There were enough cases where the diagnostic appeared to increase or not be close to converging that I increased the factor for the number of events from 3 to 5.
The factor referenced is the one that controls how likely we are to sample all combinations of species and environment in the coupon collecting problem.

It is not obvious to me whether this should have the spurious correlation removed as well, which might reduce the diagnostic further.