Merge with Research Server XI.
GenerateCases was updated to accomodate the "Inf" distance case, but see the previous commits for what happened as a result of some other problems.
Essentially, the documentation is slightly wrong here: line 121 should be moved to be after "h." and Inf should now be last in the vector on line 131 to accurately reflect what happened.

GeneratePMEsParallel was fixed to work better and to make use of arrival and extinction functions that do not permit "running out" of events.
The "cost" of these functions is that we lose control over the exact number of events we receive.

I also note that I evidently set Const to 7 rather than 5 as noted in a previous commit.