Merge with Research Server XII.
With all of the aforementioned having been discussed, what is there left?
The second attempt at simulation, which parallelises and sets up the foundation for later simulations, includes parameter variation built into the code and usage of the updated arrival and extinction distribution functions.

This is accompanied by a new gallery, patterned after the 3rd version in a previous commit.