Heterogeneity Decisions
So a couple of decisions have been made. Figure 3 is uninteresting in the case of increased heterogeneity of environments or homogeneous environments. I'm committing this, and then moving it to extraneous in recognition.

Second, it's been decided that the high heterogeneity system should be included as a separate column in figure 4 (and thus should contribute to the supplemental calculations, including supplemental figures and percentage of simulations completed). This will require some edits of the HighK6 to reflect this, but they will be committed here first.