Implementation of Hik6
Easiest way is to create duplicates of the Viking scripts. These can then be re-run without running everything else. I've then edited the combine scripts so that they take the set number + 3, which effectively translates them to a 4th set. We'll then need to add the fourth set of data, but plotting will translate easily.