Removing some Tests.
I'm removing these files because they now fail to run before the step that's really relevant (and before we can do the proper check as a result). The goal had been to find the old populations that I had not recorded and then see if the abundances obtained from the older methods were the same as those obtained via  the more up-to-date implementation. Unfortunately, the generation of the initial abundances doesn't converge it looks like (after two days running).