Elevating Mutualism to Tracking
Group thinks this path is worth investigating more thoroughly. Roughly, we have a set of derivative files from Figure 3 (since we're performing case studies and not full simulation studies). Create generates pool, matrices, and simulations. Image generates the classical Figure 3. Invasibility and Specialism generate temporal beta measures and (try to) measure the generalism-specialism trade-off that (maybe) emerges during the assembly process.

All of these files rely on Mutualism (or, in one version of the data folders not included, Mutualism2). These implement all of the details necessary to run a mutualistic system, but these need to be updated to more clearly take parameters.