Renamed Mutualism Folder, Added Time Series
Main contribution is TimeSpaceAndTimeSeries2.R, which might form the basis for the next project. It's a bit lengthy to describe, but tries to capture the differences between time series and space for time substitution when analysing an ecosystem, except we also know the true ecosystem because we have the controlled simulation(s).