Intermediate for Paired -> Average and Time-Space
Removal of if statement that could cause code not to run in 0-Functions.
Change of barrier for the bootstraps that could cause error in code (due to exceeding the data allotted).

Added consideration for the timegaps explicitly, rather than relying on filtering it out.
Addition of a new initial plot to show, that takes a gap of 10 at maximal distance. We'll then iterate this to show how it collapses for various time gaps and space gaps.

The larger coding challenge at the moment then will appear to be redoing the original averaged plot for a single time point.