Update TimeSpaceAndTimeSeries-1-Bootstraps.R
Lots of changes over the past few days.

Support for single split files. Could be extended with either lists or with ... support.

Recognising the different pool formats.

Found a slight optimization and use for knowing the time differences outside of the main foreach.

Added in true (instead of sampled) beta diversity plot.