Update TimeSpaceAndTimeSeries-2-AnalysisAlphaGamma.R
As previous commit.

Employing the timediffs to select a certain time gap.

Added in support for and implementation of beta diversity ribbon plots. Note the use of a new function, mentioned in the next commit.

Added in minimal, but not good, support for logarithmic time scales. I'm not sure if "nothing" is happening (I mean, it is, so maybe the stochastic changes are averaging out?) or if the implementation of logarithmic time scales is insufficient.