Update TimeSpaceAndTimeSeries-0-Functions.R
As before.

To make the TS&TS-1 Intervention version read and work better, I've stripped out the loading of the results into a separate function.

To support the Beta implementation in TS&TS-2 AlphaBetaGamma, I've added an uniform way to addDistanceColumns. This is only used in Beta, but should be rolled out to Alpha and Gamma once I'm confident.

I've also implemented a function that converts the sampling strings (two forms!) into vegan::vegdist's. This should be general purpose enough to see a long use life.