Create TimeSpaceAndTimeSeries-1-Intervention.R
Beginning to work on (but note the need to implement the actual workings) a variant of TS&TS-1 bootstraps with an intervention. As I go, I'm hoping to see about functionalizing some of the code.

Bootstraps isn't deprecated yet, and I'm not sure if I should move towards deprecation, but this file is aiming to essentially do the same thing if timeInterventionRandom, patchInterventionRandom, and relabelPoolIntervention are FALSE. These extensions should create various forms of intervention.