Beginning to implement the intervention
Initial attempt is to implement it within the Sample function, but that doesn't seem right. The problem with doing it entirely there is that the consistency between the two methods is lost (i.e., additional variation is introduced that I don't think we want). I think I want to instead make the sampling dataframe more responsible.