First attempt at non-null intervention done
We're trusting the rowwise substitute we have. Otherwise, this entails essentially just swapping when we sample from the intervention set and when we sample from the base set. This is cheap because all of the substitutions' scaffolding happens inside the sampling dataframe now.

Note that we should refactor the language used, as it is beginning to get a bit clunky (when use "intervention vs alternate" and "actual vs base".