Update TimeSpaceAndTimeSeries-1-Intervention.R
We've run into a bit of an issue regarding the characteristic times. Effectively, the intervention time scale is fundamentally different than the base time scale. As a result, I'm going to need to instead put everything on the same time scale using the reaction times. Not a perfect solution (since the reaction times are approximations) but the best one I've got.

Also dplyr::bind_cols was reformatting as tibbles which was both slower and not the format we wanted. Changed to cbinds just to be safe (since tibble's renames might cause problems).