Adding Bray-Curtis
Given the addition of another Beta diversity measure, I decided to functionalise the process of converting the prepared data frame into beta diversity (to reduce copy-pasting -> risk of bugs). I've checked that the old bootstrapSamplesPairedBeta is equivalent to bootstrapSamplesPairedBeta_Jaccard generated by the new function. If the same works for Timed, then I'll move the new function into the 0-Functions.R file (which will help reduce the number of lines in the script).