Reverting initial attempt at "True" values
These should probably be based on the control time series only, else it creates a false image of space for time as being equally true. We are explicitly trying to capture the sources of error due to sampling (time series vs factual) and due to substitution (space for time vs factual - sampling error) in addition to the otherwise omnipresent and unavoidable causal error (factual vs counterfactual).