Matrix Interpolation Function
As described in-line, we pick a function that interpolates between two matrices that has a balance between how much transition occurs in the middle and how much the interpolation occurs at the edges.

I've also suggested a form for applying the intervention, but details are unspecified at current because we need time dependence inside the per capita dynamics (both for if we decide to have a pool interpolation and for the matrix interpolation). This should be in the main trunk and then merged in here.