Updating Per Capita Dynamics Type 1 Factory
Now allowing for t, y, and parms to be passed through if the reproduction rate or interaction matrix are functions.

We *don't* do the same for NumEnvironments, because that necessitates a larger rewrite that should be thought about deeply before proceeding.

I've also added some checks to make sure that the new functionality works.

Thinking about it, I may want to make sure that they are conformable, but the error handling down the line should be sufficient, so I'm a bit torn. I'll leave it out for now.