Update TimeSpaceAndTimeSeries-1-Simulation.R
Edits that technically implement things, but to an incomplete spec (hence lines 408 and 409).

Path forward is clear, but we're not finished yet. Might be able to get a trial run done before Christmas. We'll then have a week to finish up and get some trial simulations done. Don't forget the cleanup either.