Update TimeSpaceAndTimeSeries-1a-Simulation.R
Caught a problem with the switchtime and timespan; these need to be on the evaluation time scale rather than on the characteristic time scale (which is used for comparison).

Additionally, fixed a problem regarding the types. I had to lose a feature (multiple switchtimes and timespans), but we now make sure that it is function(blockdiagonal(matrices)) rather than blockdiagonal(matrices(function)). A function or static matrix/vector is required by  the dynamics functions and there is no implemented interpretation of a blockdiagonal of matrix (list) of functions.