Create UpdateTruncNormBounds.R
I was asked to implement a way to sample a subset of the truncated normal distributions. My solution is to move the boundaries of the truncated normal (with underlying Normal(mu, sigma)) from [a, b] to new locations. Then we can just draw from truncated normal distributions like, well, normal.

Next step is to integrate this function so that we can use it in generating interaction matrices.