Version Bump + Fix for New Parameters
Version bump because of changes to the main functions in the R folder.

Fix for new parameters should allow for more general parameters, but required use of Bing Chat in order to find a solution for converting a list of arguments into something parseable by purrr's partialize. Additional credit goes to stackoverflow.com/a/47012149 for providing a function to convert a character string of arguments into a list of arguments to partialize with.

And, of course, this was motivated by the attempt to implement treatments with different parameters than were held in the Parameters argument of our Interaction Matrix generator function.