Finishing Building Sampling
Applied some corrections for changes in format, parallelization, a parameter dependent createSamplingTimes that iterates previous design, and the architecture for the actual sampling. Note the filename format and contents have been decided as well (lines 115 - 118, with all runs samples as on lines 204 forward).

In the process, some edits have been made to the simulations regarding pass throughs (to facilitate plotting) and a new version of the sampling function to fix a bug regarding not specifying the pool.

We're now moving onto plotting implementation.