Update MNA-Image-ExampleOutcome-Create.R
Huge number of changes. Enough I think to justify creating a new file. I'm going to finish with what I've got at the moment and then perform the initial refactor afterwards.

Note esp. the introduction of an intervention here, which was effectively a no-no until now. I'll talk with the others about refactoring. In principle, I'd like to do a mass refactor, but I don't think that's in the cards.