Moving older files into a deprecated folder
We're not using them at the moment and are unlikely to return to them. I don't plan on deleting any of these files at the moment, but might do so if we decide strongly to travel in the current direction. (Note: we may want to be picky about which files to delete, and which to move to the extraneous area. E.g., mutualism implementations.)