Update TimeSpaceAndTimeSeries-6d-PlotDiversities.R
Lots of changes to presences that were required to make the plots work. 293 - 330 or so are trivial. 333 - 343 take the median abundance over a time period and then delete abundances below the (TECH DEBT: HARD CODED) elimination threshold. 356 - 386 allow me to transfer the diversity affinity groupings over to the presences, at the cost of the obvious dependency. Then the actual plot, which required some guesswork to get working in a minimally meaningful way.