Testing calculateAbundanceMetrics
Okay, so good news and bad news.

Good news: I can reproduce the results of my function "by hand" as it were.

Bad news: there was a "spatial" bug that is probably propagated from the earlier CalculateDiversities family of functions in which I garbled the labels for the environments. I need to go back and check to see what this affects, but I'm optimistic that it won't be much due to the aggregations we were performing and that the error should be consistent (since the ordering that caused it and number of environments should be consistent and the error has nothing to do with the temporal axis).

Ugly News: I have no idea why it isn't matching our intuitions about when the distances should be high/low.