Create VisualisingFoodWebs.R
I've not committed the setwd(...) command or the modules used.

Okay, so here is a file that is overdue for committing. This file loads the results from an intervention, the relevant underlying data, and the diversity analyses of species identity and abundance. We try to make sure things have shared formats, break the data collected by patch (to facilitate plotting the food webs) and try to reconstruct the per-capita effects on the population at a given point in time. These naturally inform the edges between nodes which are time resolved in the animation::saveVideo loop. (Note the dangerous bug, fixed in an earlier version, of forgetting to actually print the plot!)

Node data tries to incorporate the properties of a node at a given time, edges represent 2-body interactions. To capture the intraspecific 2-body interactions, we add false nodes that have currently been turned off. Dispersal is present but also always turned off.

Thresholding of edges is permitted but turned off. It was replaced with alpha scaling due to the magnitude of the effect on a percentage scale. Fill of a node (species) reflects the per capita change. Abundance is reflected by the size of the border.

One controversial bit was the usage of bar plots inside of nodes. This proved to be too much information for reading the plots.

Multiple layouts have been attempted, but the one with the most spread has been an undirected backbone with the vertical position replaced with the logarithmic size. I'd like to do better, but haven't found a solution.

This file creates a video for visualising food webs on (currently two) patches as time progresses.
Primary TODO is to make sure the pre/no intervention case is also implemented.

Other TODOs include making sure it generalises to 10, investigating whether a shiny app would be better (a single image takes a few seconds to generate!), fixing the fill (more on that in a moment), and making some visual adjustments. It also might be good to create a row of summary statistic panels through time with a moving vline to indicate the current time.