Create temp_DoubleCheck.R
With the aforementioned commit's text having been said, the main caveat at the moment is trying to figure out why my calculation of the per capita growth rate of a species differs from using the existing functions. This script can be run after the VisualisingFoodWebs.R script and reports the values as if they were calculated within the usual dynamics functions (unless, of course, I've made an error here instead!).

I trust this file over the VisualisingFoodWebs.R file due to 1) simplicity and 2) better agreement with what is happening in the resulting videos.  On the time step I've attempted this on, the population was growing, but the reported per capita change in abundance was negative in the video. This file, on the other hand, reports a positive change. This suggests that it was not a plotting failure, but rather a computational one.