Update TimeSpaceAndTimeSeries-6b-Intervention.R
Discovered bug that was retrieving the previous function for adjusting intrinsic growth and decay rates rather than the current function. This means that each mismatched call has not invoked the right function (and is thus wrong and matches with the file that has the new function as the same as the shared old function).