Update VisualisingFoodWebs.R
Moved quite a bit into helper functions, since I need to call them twice in order to facilitate the construction of the layout. Code appears to work on a small test.

(More could be moved out to better the engineering of the code, but I'm not at a refactoring stage.)

Also reduced the number of prints to try to get more useful information out of the highest effect size (fill).