Create TimeSpaceAndTimeSeries-6f-SamplingPlots.R
Adjusting 5daside as 6f. First I'm going back through the code to make sure it matches with the new use cases. Then, I need to go back through and repurpose it to start showing how the sampling is working in comparison to the factual reality and the counterfactually determined influence of the intervention.