Double-checking Sampling Code
Adding comments and going through an example to make sure everything works. Found one improvement that needed to be made to sampleFromResults2 to reflect that we want to use it for more than just Basal vs Consumer now. This means making the types argument more generic.

I've also changed where we compare before and after the intervention so that it is some form of physical time rather than computational time, but I'm not satisfied. The interventions are initiated to the time just before the intervention occurs, but that time can change from intervention row - 1 to potentially earlier computational steps. And, of course, this means that converting to physical times becomes harder as well.  I don't think what I've done is anything more than a bandage. I'd actually want to take the sample from before the intervention *file* to the intervention files, but I'm not sure if that breaks some of the counterfactual approach.