Update TimeSpaceAndTimeSeries-6f3-WindowPlots.R
Adding space-for-time plot equivalents. Note that the match is quite strong between the 3 versions, so it is overkill. The match is probably so strong because of the linearity of the operations, and I thus don't expect this to be the case for all methods of comparison.

Otherwise, the difference between slope variations for space-for-time and time series looks consistent. The space-for-time implementation of differencing seems less obvious to interpret than the one for time series though.