Create transitionsMatrixIdea.R
While the idea works, the problem we have is that there is uneven sampling. The idea had been to convert to a continuous time Markov chain, then to a discrete time Markov chain to know the step sizes that we can use in our system. Hence, we need the thing we're looking for in effect!