Update SketchComparisons.R
Determined there was a bug with the way I was assigning time to the case of ParentRun, "After". Since my interventions were shifted to start at 0, the ParentRun, "After" case was being evaluated at the ParentRun, "Before" times. The InterventionRun cases were shifted appropriately (lines 38 - 46), and the ParentRun, "Before" cases are unaffected (because there is no shift).

I've added a pair of new columns that reverse the shift for the ParentRun, "After" case specifically, and reduced the number of evaluations of the abundance functions by half. I've also separated out the simulations so that we can see clearer striping in the plots, rather than the noise of different color stripes overlaid on each other.