Update SketchComparisons.R
Saving individual plots + adding new plot and new comparisons suggested by Inês. Need to rerun to create plots (and then make sure to save them) and to verify that the new plots and comparisons work. Also need to likely explicitly call the new comparisons for some of the plots.