Create v7 Simulate as Function
I *think* instantiate PerCapitaDynamics could reasonably be turned into a separate function. There is some amount of control flow and complexity here that I don't think is happening with the construction of the other objects.

Otherwise, we should be able to just call the simulate wrapper with the appropriate dictionary and seed choices, instead of having to make a large number of scripts or script options to handle it. Querying as to the appropriate choice can be handled with the Dictionaries objects, which I've now added IDs to.