Fixing problems in 7c/d
Problem with the ID handler in that the baseTag is different from the dirTag, when I had thought they were the same. Need to fix ID handler, but I'm using the work around I already programmed in.

Additionally, poor placement of the appendID and a bad x_properties that was accidentally retained.